backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
psycopg2-binary==2.9.7
alembic==1.11.1
python-multipart==0.0.6
pydantic==1.10.9
celery==5.3.1
redis==4.5.5
aiofiles==23.1.0
python-dotenv==1.0.0
websockets==11.0.3
httpx==0.24.1
python-jose[cryptography]==3.11.0
passlib[bcrypt]==1.7.5
pytest==7.4.0
